feat: 认证侧 SDK
