+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './ManagementClient';
 export * from './ManagementClientOptions';
-export * from './AuthClient';
-export * from './AuthClientOptions';
-export * from './AuthClientInterface';
+export * from './AuthenticationClient';
+export * from './AuthenticationClientOptions';
+export * from './AuthenticationClientInterface';
 export * from './RsaPublicKey';
 export * as Models from './models';
